[ feat ] 해시태그 컴포넌트 분리
